+++ RandomBot/Models/HocModel.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace RandomBot.Models
+{
+    public class HocModel
+    {
+        public string Name { get; set; }
+        public string Chip { get; set; }
+        public string Lethality { get; set; }
+        public string Pierce { get; set; }
+        public string Precision { get; set; }
+        public string Reload { get; set; }
+        public string Range { get; set; }
+        public string NormalAttack { get; set; }
+        public string Skill1 { get; set; }
+        public string Skill2 { get; set; }
+        public string Skill3 { get; set; }
+    }
+}