+++ RandomBot/Modules/InfoModule/InfoModule.cs
@@ -12,6 +12,11 @@ namespace RandomBot.Modules.InfoModule
         {
             var embed = new EmbedBuilder()
                 .WithTitle("Development Status")
+                .AddField("Build 0.4.0", @"
+Added $hoc command
+Added $whale command
+Added $scam command
+Improved database access related task")
                 .AddField("Build 0.3.7", @"
 Added $Dorime command
 Updated Entities")
@@ -21,14 +26,6 @@ Changes on reminder module to improve code readability
 Message for $delete command changed
 Fixed message on $timeconvert command
 Improved code readability on bot startup")
-                .AddField("Build 0.3.5", @"
-Improved Reminder Module
-Changes Gunfu ROF calculation from 30 FPS to 60 FPS cap")
-                .AddField("Build 0.3.4", @"
-Added text edit on image feature
-Added Freedom Module
-Added precise search feature for GunfuService.cs
-Improved flexibility on adding Reminder")
                 .AddField("More?", @"Details on https://gitlab.com/adrianch/RandomBot/blob/dev-0.4.0/CHANGELOG.md")
                 .WithColor(Discord.Color.DarkRed);
             await ReplyAsync("", embed: embed.Build());